envoy: Bump envoy image to the latest v1.26.x
[ upstream commit 34415bd203fb5cfe32fe829f04cd691ff6d35712 ]

This commit is to perform the below:

- Update envoy image to latest build from v1.26 branch
- Include new resources into pkg/envoy/resource for serialization
  - grpc related resources are for upcoming support with GRPCRoute
- Change policy id to uint32, related to https://github.com/cilium/proxy/commit/f37daf70991e4d58fc610244fb559cf48ea475e2

Related build: https://github.com/cilium/proxy/actions/runs/7070197416/job/19246669763

+++ pkg/envoy/server_test.go
@@ -259,7 +259,7 @@ var ExpectedHttpRule122HeaderMatch = &cilium.PortNetworkPolicyRule_HttpRules{
 }
 
 var ExpectedPortNetworkPolicyRule12 = &cilium.PortNetworkPolicyRule{
-	RemotePolicies: []uint64{1001, 1002},
+	RemotePolicies: []uint32{1001, 1002},
 	L7:             ExpectedHttpRule12,
 }
 
@@ -268,7 +268,7 @@ var ExpectedPortNetworkPolicyRule12Wildcard = &cilium.PortNetworkPolicyRule{
 }
 
 var ExpectedPortNetworkPolicyRule122HeaderMatch = &cilium.PortNetworkPolicyRule{
-	RemotePolicies: []uint64{1001, 1002},
+	RemotePolicies: []uint32{1001, 1002},
 	L7:             ExpectedHttpRule122HeaderMatch,
 }
 
@@ -277,7 +277,7 @@ var ExpectedPortNetworkPolicyRule122HeaderMatchWildcard = &cilium.PortNetworkPol
 }
 
 var ExpectedPortNetworkPolicyRule1 = &cilium.PortNetworkPolicyRule{
-	RemotePolicies: []uint64{1001, 1003},
+	RemotePolicies: []uint32{1001, 1003},
 	L7:             ExpectedHttpRule1,
 }
 
@@ -426,10 +426,10 @@ var ExpectedPerPortPolicies12RequiresV2 = []*cilium.PortNetworkPolicy{
 		Port:     80,
 		Protocol: envoy_config_core.SocketAddress_TCP,
 		Rules: []*cilium.PortNetworkPolicyRule{{
-			RemotePolicies: []uint64{1001, 1002},
+			RemotePolicies: []uint32{1001, 1002},
 			L7:             ExpectedHttpRule1,
 		}, {
-			RemotePolicies: []uint64{1002},
+			RemotePolicies: []uint32{1002},
 			L7:             ExpectedHttpRule12,
 		}},
 	},
